Add pull request verification (#19)
<!--
> [!Tip]
> This pull request is best reviewed commit-by-commit.
-->

<!--
Uncomment the section above if you used granular commits to compose this
change.
-->

## Context

<!--
Provide any context you think a reader should know going into this pull
request.

For example, if this change is intended to fix a bug, describe how the
bug was discovered and / or how to replicate it.

If this is a new feature, describe the domain problem this feature
addresses.
-->

We are just starting this project up and want to ensure that we do not
introduce regressions as we make changes.

## Intention <!-- (optional) -->

<!--
Feel free to delete this section if you feel like it does not add any
value given the information laid out in the "Context" section.

This can describe the desired behavior of the feature you are
implementing.
-->

Introduce an introductory pull request verification template.

## Implementation

<!--
Describe the technical implementation details of this change.
-->

- Create a reusable action for install .NET
- Create a reusable action for install pnpm
- Create a pull request workflow
  - Run build
  - Run lint
- Add recommended extensions for creating workflows and for having live
linting
- Set default tab size to 2
- Improve build caching for .NET API
- Fix a warning discovered by the linter
- Ensure the linter throws an error if any problems are found when
running from the command line
- Clean up some crusty turbo tasks.

## Testing

<!--
How did you test this change?

Did you add unit tests? End-to-end tests?

If necessary, paste screenshots or demo videos in this section.
-->

The linter
[discovered](https://github.com/computost/teachutost/actions/runs/14433119937/job/40470318915#step:4:29)
a warning we had previously ignored. It needed to be fixed in order for
this workflow to pass. In my mind, that feels like a pretty good test.

## Notes

<!--
If there are any details about this pull request that you do not feel
fit into any other sections, add them here.
-->

This is only a first pass at pull request verification. This template
will grow significantly as we add more features. Here are some things
I'd like to add:

- Test multiple operating systems for the build
- Leverage remote caching?

## Ticket

Resolves #7

<!--
Add the ticket number to the statement above.

This uses a keyword to automatically take action on the issue when the
pull request merges.


https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword

Feel free to modify the keyword if another is more appropriate. For
example: "Fixes."
-->

## Wholesome Toll

<!--
Feel free to share anything here that makes you happy.

It can be a poem, a hot take, a family photo, or whatever you like.

Maybe you'll make a future reader's day just a little bit better.
-->

Broke: vibe coding

Woke: Ricky Sparkle coding


![PXL_20250131_225320340.jpg](https://github.com/user-attachments/assets/ab18d7fc-a89d-4265-b94e-f032cb4b30cb)