Create a pull request template (#18)
> [!Note]
> The quotation blocks below represent comments in the template. They
have been converted into quotation blocks to improve the readability of
this pull request.

> > [!Tip]
> > This pull request is best reviewed commit-by-commit.

> Uncomment the section above if you used granular commits to compose
this change.

## Context

> Provide any context you think a reader should know going into this
pull request.
>
> For example, if this change is intended to fix a bug, describe how the
bug was discovered and / or how to replicate it.
>
> If this is a new feature, describe the domain problem this feature
addresses.

As a highly collaborative team setting out to build a large-scale
application, it is important that we are able to review the history of
changes made and understand the intent behind those changes.

## Intention `(optional)`

> Feel free to delete this section if you feel like it does not add any
value given the information laid out in the "Context" section.
>
> This can describe the desired behavior of the feature you are
implementing.

Introduce a [pull request
template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository)
that can be used as a starting checklist to guide contributors through
the information we might need for both reviewing pull requests and for
building confidence in future changes.

## Implementation

> Describe the technical implementation details of this change.

Added these sections to `.github/pull_request_template.md`.

## Testing

> How did you test this change?
>
> Did you add unit tests? End-to-end tests?
>
> If necessary, paste screenshots or demo videos in this section.

This pull request _is_ the test. As a reviewer, consider whether this
will be able to meet the needs of the team going forward as we build
more features for the application.

It doesn't need to be perfect in the first attempt, but hopefully this
is enough to get us moving.

> ## Notes

> If there are any details about this pull request that you do not feel
fit into any other sections, add them here.

## Ticket

Resolves #1.

> Add the ticket number to the statement above.
>
> This uses a keyword to automatically take action on the issue when the
pull request merges.
>
>
https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
>
> Feel free to modify the keyword if another is more appropriate. For
example: "Fixes."

## Wholesome Toll

> [!Note]
> The header above is commented out by default. It has been uncommented
in this pull request.

> Feel free to share anything here that makes you happy.
>
> It can be a poem, a hot take, a family photo, or whatever you like.
>
> Maybe you'll make a future reader's day just a little bit better.

Taken at [Rhododendron Species Botanical
Garden](https://maps.app.goo.gl/GRFxUgUkKMv1d7QNA)


![PXL_20250404_195702543.jpg](https://github.com/user-attachments/assets/9ac34121-5c8d-42bc-b620-7a385e31ede1)