Expose API over local network (#16)
## Context

Currently, we are able to test the application locally on a development
machine. However, because everything is hard-coded to point to
`localhost`, we cannot test it on mobile devices.

## Intention

Allow an engineer to open and test the app from their mobile device.

## Implementation

- Modify the API to accept connections from any origin while running in
`Development` mode.
- This requires that we set the host name of the API to `0.0.0.0`, which
is not a valid address on Windows machines. To prevent the application
from opening a browser on a non-responsive address, the API no longer
opens the browser on startup. It must be opened manually.
- Modify the mobile app to calculate the address of the API based on the
address of the host.
- **Note:** Eventually, we should figure out how to do this by passing
an environment variable into the mobile app. This is complicated because
the environment variable needs to get the local IP address of the host
machine, which is no simple task; a machine can have multiple network
cards, and therefore multiple IP addresses. We need to figure out which
one to use.

## Testing

I tested this manually for now. As we introduce more testing suites, we
can add automated tests.

Here is a screenshot of the mobile app from my phone.


![Screenshot_20250413-102120.png](https://github.com/user-attachments/assets/a091a810-ef8a-41a1-9bd2-22308a00cd16)

## Ticket

This pull request resolves issue #15.