21 run dev environment in a container (#22)
## Context

Inconsistent dev environments are causing issues for DevX

## Implementation

Set up devcontainers to containerize the dev environment

## Testing

Built and ran the devcontainer with devpod. Executed turbo scripts to
restore dev tools, build the application, and run the application in dev
mode. Checked that the API and websockets are working on local and on
mobile

## Notes

@melody-universe is working on a Readme PR which will contain
documentation on setting up and using devpod

## Ticket

Resolves #21